Improved support for one-click unsubscribe for notifications.
FossilOrigin-Name: bca95cbf9c2bf230efe14793f0389db34c6636d3b44b0863161a6997a6d84dcb

+++ src/alerts.c
@@ -49,6 +49,7 @@ static const char zAlertInit[] =
 @ --     a - Announcements
 @ --     c - Check-ins
 @ --     f - Forum posts
+@ --     k - ** Special: Unsubscribed using /oneclickunsub
 @ --     n - New forum threads
 @ --     r - Replies to my own forum posts
 @ --     t - Ticket changes
@@ -88,7 +89,7 @@ static const char zAlertInit[] =
 @   sentDigest BOOLEAN DEFAULT false, -- digest alert sent
 @   sentMod BOOLEAN DEFAULT false     -- pending moderation alert sent
 @ ) WITHOUT ROWID;
-@
+@ 
 @ -- Obsolete table.  No longer used.
 @ DROP TABLE IF EXISTS repository.alert_bounce;
 ;
@@ -877,7 +878,7 @@ void email_header_to(Blob *pMsg, int *pnTo, char ***pazTo){
   Blob v;
   char *z, *zAddr;
   int i;
-
+  
   email_header_value(pMsg, "to", &v);
   z = blob_str(&v);
   for(i=0; z[i]; i++){
@@ -891,7 +892,7 @@ void email_header_to(Blob *pMsg, int *pnTo, char ***pazTo){
 }
 
 /*
-** Free a list of To addresses obtained from a prior call to
+** Free a list of To addresses obtained from a prior call to 
 ** email_header_to()
 */
 void email_header_to_free(int nTo, char **azTo){
@@ -917,7 +918,7 @@ void email_header_to_free(int nTo, char **azTo){
 **     Content-Transfer-Encoding:
 **     MIME-Version:
 **     Sender:
-**
+**     
 ** The caller maintains ownership of the input Blobs.  This routine will
 ** read the Blobs and send them onward to the email system, but it will
 ** not free them.
@@ -930,7 +931,7 @@ void email_header_to_free(int nTo, char **azTo){
 ** email address based on a hash of zFromName and the domain of email-self,
 ** and an additional "Sender:" field is inserted with the email-self
 ** address.  Downstream software might use the Sender header to set
-** the envelope-from address of the email.  If zFromName is a NULL pointer,
+** the envelope-from address of the email.  If zFromName is a NULL pointer, 
 ** then the "From:" is set to the email-self value and Sender is
 ** omitted.
 */
@@ -1047,7 +1048,7 @@ void alert_send(
 */
 /*
 ** SETTING: email-admin               width=40
-** This is the email address for the human administrator for the system.
+** This is the email address for the human administrator for the system. 
 ** Abuse and trouble reports and password reset requests are send here.
 */
 /*
@@ -1082,7 +1083,7 @@ void alert_send(
 ** last batch of "your subscription is about to expire" emails were
 ** sent out.
 **
-** email-renew-cutoff is normally 7 days behind email-renew-warning.
+** email-renew-cutoff is normally 7 days behind email-renew-warning.  
 */
 /*
 ** SETTING: email-send-method         width=5 default=off sensitive
@@ -1090,7 +1091,7 @@ void alert_send(
 ** "off", "relay", "pipe", "dir", "db", and "stdout".  The "off" value
 ** means no email is ever sent.  The "relay" value means emails are sent
 ** to an Mail Sending Agent using SMTP located at email-send-relayhost.
-** The "pipe" value means email messages are piped into a command
+** The "pipe" value means email messages are piped into a command 
 ** determined by the email-send-command setting. The "dir" value means
 ** emails are written to individual files in a directory determined
 ** by the email-send-dir setting.  The "db" value means that emails
@@ -1135,7 +1136,7 @@ void alert_send(
 
 /*
 ** COMMAND: alerts*
-**
+** 
 ** Usage: %fossil alerts SUBCOMMAND ARGS...
 **
 ** Subcommands:
@@ -1744,7 +1745,7 @@ void subscribe_page(void){
 ** by the hex value zName.  Then paint a webpage that explains that
 ** the entry has been removed.
 */
-static void alert_unsubscribe(int sid){
+static void alert_unsubscribe(int sid, int bTotal){
   const char *zEmail = 0;
   const char *zLogin = 0;
   int uid = 0;
@@ -1761,9 +1762,20 @@ static void alert_unsubscribe(int sid){
     style_header("Unsubscribe Fail");
     @ <p>Unable to locate a subscriber with the requested key</p>
   }else{
-    db_multi_exec(
-      "DELETE FROM subscriber WHERE subscriberId=%d", sid
-    );
+    db_unprotect(PROTECT_READONLY);
+    if( bTotal ){
+      /* Completely delete the subscriber */
+      db_multi_exec(
+        "DELETE FROM subscriber WHERE subscriberId=%d", sid
+      );
+    }else{
+      /* Keep the subscriber, but turn off all notifications */
+      db_multi_exec(
+        "UPDATE subscriber SET ssub='k', mtime=now() WHERE subscriberId=%d",
+        sid
+      );
+    }
+    db_protect_pop();
     style_header("Unsubscribed");
     @ <p>The "%h(zEmail)" email address has been unsubscribed from all
     @ notifications.  All subscription records for "%h(zEmail)" have
@@ -1795,7 +1807,7 @@ static void alert_unsubscribe(int sid){
 **    *    The sid= query parameter contains an integer subscriberId.
 **         This only works for the administrator.  It allows the
 **         administrator to edit any subscription.
-**
+**         
 **    *    The user is logged into an account other than "nobody" or
 **         "anonymous".  In that case the notification settings
 **         associated with that account can be edited without needing
@@ -1925,9 +1937,9 @@ void alert_page(void){
       zErr = mprintf("Select this checkbox and press \"Unsubscribe\" again to"
                      " unsubscribe");
     }else{
-      alert_unsubscribe(sid);
+      alert_unsubscribe(sid, 1);
       db_commit_transaction();
-      return;
+      return; 
     }
   }
   style_set_current_feature("alerts");
@@ -2090,6 +2102,10 @@ void alert_page(void){
     @  Wiki</label>
   }
   @ </td></tr>
+  if( strchr(ssub,'k')!=0 ){
+    @ <tr><td></td><td>&nbsp;&uarr;&nbsp;
+    @ Note: User did a one-click unsubscribe</td></tr>
+  }
   @ <tr>
   @  <td class="form_label">Delivery:</td>
   @  <td><select size="1" name="sdigest">
@@ -2184,7 +2200,7 @@ void renewal_page(void){
 /* This is the message that gets sent to describe how to change
 ** or modify a subscription
 */
-static const char zUnsubMsg[] =
+static const char zUnsubMsg[] = 
 @ To changes your subscription settings at %s visit this link:
 @
 @    %s/alerts/%s
@@ -2196,6 +2212,7 @@ static const char zUnsubMsg[] =
 
 /*
 ** WEBPAGE: unsubscribe
+** WEBPAGE: oneclickunsub
 **
 ** Users visit this page to be delisted from email alerts.
 **
@@ -2208,6 +2225,9 @@ static const char zUnsubMsg[] =
 ** Non-logged-in users with no name= query parameter are invited to enter
 ** an email address to which will be sent the unsubscribe link that
 ** contains the correct subscriber code.
+**
+** The /unsubscribe page requires comfirmation.  The /oneclickunsub
+** page unsubscribes immediately without any need to confirm.
 */
 void unsubscribe_page(void){
   const char *zName = P("name");
@@ -2225,15 +2245,17 @@ void unsubscribe_page(void){
   if( zName==0 ) zName = P("scode");
 
   /* If a valid subscriber code is supplied, then either present the user
-  ** with a confirmation, or if already confirmed, unsubscribe immediately.
+  ** with a comformation, or if already confirmed, unsubscribe immediately.
   */
-  if( zName
+  if( zName 
    && (sid = db_int(0, "SELECT subscriberId FROM subscriber"
                        " WHERE subscriberCode=hextoblob(%Q)", zName))!=0
   ){
     char *zUnsubName = mprintf("confirm%04x", sid);
     if( P(zUnsubName)!=0 ){
-      alert_unsubscribe(sid);
+      alert_unsubscribe(sid, 1);
+    }else if( sqlite3_strglob("*oneclick*",g.zPath)==0 ){
+      alert_unsubscribe(sid, 0);
     }else if( P("manage")!=0 ){
       cgi_redirectf("%R/alerts/%s", zName);
     }else{
@@ -2314,7 +2336,7 @@ void unsubscribe_page(void){
     alert_sender_free(pSender);
     style_finish_page();
     return;
-  }
+  }  
 
   /* Non-logged-in users have to enter an email address to which is
   ** sent a message containing the unsubscribe link.
@@ -2723,7 +2745,7 @@ EmailEvent *alert_compute_event_text(int *pnEvent, int doDigest){
                    zSub, zTitle);
     }else{
       blob_appendf(&p->hdr, "Subject: %s %s\r\n", zSub, zTitle);
-      blob_appendf(&p->hdr, "Message-Id: <%.32s@%s>\r\n",
+      blob_appendf(&p->hdr, "Message-Id: <%.32s@%s>\r\n", 
                    zUuid, alert_hostname(zFrom));
       zIrt = db_column_text(&q, 4);
       if( zIrt && zIrt[0] ){
@@ -3150,7 +3172,7 @@ int alert_send_alerts(u32 flags){
         blob_appendf(&fhdr, "To: <%s>\r\n", zEmail);
         blob_append(&fhdr, blob_buffer(&p->hdr), blob_size(&p->hdr));
         blob_init(&fbody, blob_buffer(&p->txt), blob_size(&p->txt));
-        blob_appendf(&fhdr, "List-Unsubscribe: <%s/unsubscribe/%s>\r\n",
+        blob_appendf(&fhdr, "List-Unsubscribe: <%s/oneclickunsub/%s>\r\n",
                      zUrl, zCode);
         blob_appendf(&fhdr,
                    "List-Unsubscribe-Post: List-Unsubscribe=One-Click\r\n");
@@ -3180,7 +3202,7 @@ int alert_send_alerts(u32 flags){
       }
     }
     if( nHit==0 ) continue;
-    blob_appendf(&hdr, "List-Unsubscribe: <%s/unsubscribe/%s>\r\n",
+    blob_appendf(&hdr, "List-Unsubscribe: <%s/oneclickunsub/%s>\r\n",
          zUrl, zCode);
     blob_appendf(&hdr, "List-Unsubscribe-Post: List-Unsubscribe=One-Click\r\n");
     blob_appendf(&body,"\n-- \nSubscription info: %s/alerts/%s\n",
@@ -3232,7 +3254,7 @@ send_alert_expiration_warnings:
         Blob hdr, body;
         blob_init(&hdr, 0, 0);
         blob_init(&body, 0, 0);
-        alert_renewal_msg(&hdr, &body,
+        alert_renewal_msg(&hdr, &body, 
            db_column_text(&q,0),
            db_column_int(&q,1),
            db_column_text(&q,2),
@@ -3302,7 +3324,7 @@ void contact_admin_page(void){
     style_finish_page();
     return;
   }
-  if( P("submit")!=0
+  if( P("submit")!=0 
    && P("subject")!=0
    && P("msg")!=0
    && P("from")!=0
